requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
openai==1.3.5
gunicorn==21.2.0
